% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_midi_to_r_object.R
\name{convert_midi_to_r_object}
\alias{convert_midi_to_r_object}
\title{Convert MIDI file to R object}
\usage{
convert_midi_to_r_object(
  midi_file = "data-raw/midi/wikisource-contrapunctus-subject.midi"
)
}
\arguments{
\item{midi_file}{A string representing the location of the MIDI file. Defaults to 'data-raw/midi/wikisource-contrapunctus-subject.midi'.}
}
\value{
A \code{MidiFramer} object representing the MIDI data.
}
\description{
This function reads a MIDI file and converts it to an R object using the \code{MidiFramer} class from the \code{pyramidi} package.
}
\section{Development}{

The \code{tuneR} package should be explored for extracting MIDI data as we are currently only using \code{pyramidi} for this step.
}

\section{More MIDI}{
 A source for more MIDI files can be found \href{https://www.bachcentral.com/midiindexcomplete.html}{here}.
}

\references{
The default MIDI file is from \url{https://en.wikipedia.org/wiki/The_Art_of_Fugue}.
}
